Move core project files to /app for consistency
