+++ src/app/layout/App.js
@@ -4,7 +4,7 @@ import {
     Container,
 } from "semantic-ui-react";
 
-import TabBarContainer from "./features/tabs/TabBarContainer";
+import TabBarContainer from "features/tabs/TabBarContainer";
 import UnitInfo from "features/unitInfo/UnitInfo";
 import Pilots from "features/pilots/Pilots";
 import Mechs from "features/mechs/Mechs";
