+++ src/index.js
@@ -4,7 +4,7 @@ import {Provider} from "react-redux";
 
 import "semantic-ui-css/semantic.css";
 
-import configureStore from "./store/configureStore";
+import configureStore from "app/store/configureStore";
 const store = configureStore();
 
 // Save a reference to the root element for reuse
@@ -13,7 +13,7 @@ const rootEl = document.getElementById("root");
 // Create a reusable render method that we can call more than once
 let render = () => {
     // Dynamically import our main App component, and render it
-    const App = require("./App").default;
+    const App = require("./app/layout/App").default;
 
     ReactDOM.render(
         <Provider store={store}>
@@ -49,7 +49,7 @@ if(module.hot) {
 
     // Whenever the App component file or one of its dependencies
     // is changed, re-import the updated component and re-render it
-    module.hot.accept("./App", () => {
+    module.hot.accept("./app/layout/App", () => {
         setTimeout(render);
     });
 }
